Renamed extract subdir to workload_manager.
